requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-core
langgraph
pydantic
python-dotenv
google-search-results